Make the Witwatersand Gold Rush less annoying.
If a human player ever had the unfortunate luck of bordering the
Witwatersand Gold Rush province, they were beset by regular event
notifications (~2 per year). This curtails the number of notifications
somewhat, while also adding an option for human players to never be
notified again.

This does not change the MTTH so as to not affect the promptness of the
2nd Boer War firing.