Enable historical DAH flag. (#6)
Part of the unused flag work (#4).

Use the historical flag (featuring a hand-drawn elephant) for an
absolute monarchy. This is the flag used in EU4 and most often seen on
websites dealing in history/vexillology.

This changes to the current monarchy flag (featuring an engraving-style
elephant) for a semi-constitutional or constitutional monarchy. (This
flag is possibly a more fanciful recreation of the historical one.)

There is literally no good reason to do this other than:

- it's kinda funny to see a drawing turn into an engraving after
  Westernisation and liberalisation
- I will do anything to add more elephant flags to the game