Platine War updates
Changed modifiers to make the Platine War have more parity and especially to make it more nearly historical. "north_coalition" modifies the ARC and its name refers to the coalition of ENT and nearby countries. The "big army" was also adjusted a few times during testing, but eventually I think I returned it to near or equal to its original condition.