Hours of editing because I'm a dingus
Considerable bugfixes to syntax. Remaining anomalies include a coded requirement to force Corrientes to retain its provinces if made a puppet under ARC. Ideally, Corrientes should be retained in its current state as a puppet instead of my brute-force re-establishment of the 3-province Corrientes, but an ideal solution awaits further research of the problem. This build should be stable through an ENT victory in the Platine War and the 1852 Constitution.