Large update, many bugfixes
Bugfixes to SBA's events. Many other bugfixes to a large number of events. Implemented the 1880 rebellion. Implemented a centennial event that is not yet properly tested but works when console-triggered. Liberal Revolutions triggers the modifier to simulate Civil War outbreaks. GMU Koldewey has a minor fix to reference of localisation.