Tuning adjustments to historical events
Things were going wrong if the Platine War or SBA Civil War happened too early or too late. Adds the treaty to hopefully keep Brazil from bumming Argentina every game. Chile still may do this, and so more testing will be required to see if it is a problem or not.