Corrientes puppetry bug again
Corrientes yet again exhibited the province bug under certain conditions after the Corrientes Revolt, which yet again has proven to be extremely resistant to fixes. This changes the fix to be done by an ai decision. Now with an added statement in the UTI exchange to make 2356 the capital again.