Italia Irredenta hopefully fixed
This complex decision becomes spammy until Italy achieves the situation where the core-addition effect will work for all provinces. The point of this decision is to add Italian cores to neighboring provinces that possess Italian pops but belong to other countries. Becuase the `allow` step was very simple and permissive, the ai tended to spam it as quickly as possible for many in-game years or until the end of the game. Whenever I tag-switched to Italy to see what was the matter, this decision was one among many on Italy's list. It is therefore not clear how many deleterious effects this spamming has on the game, but at least it makes the log frustrating to read, and it may reduce performance. I am not sure if ai Italy was prevented from using the other decisions on its list or not while spamming this one, but this appeared to be a possibility.

In the proposed fix, I tried to balance the `effect` with the `allow`, such that it can only be clicked when an actual effect can happen. Some additional changes were made during bug-testing, because it took a few tries to get it working: I learned some things about decisions and about what scopes are allowed in them.