Fix for Trient and Triest
Things seem to have worked for Aosta and Savoy, and now the decision is fixed to stop being spammy when Trient and Triest are not yet claimed. Venice is comparably simple and apparently does not need a fix. Here I ended up needing to change more things than I had expected in order to fix the decision for Trient and Triest.