CRT rejoining ARC, infinite war fix
Corrientes was incorrectly notifying Entre Rios that it was becoming a puppet of the Argentine Confederation, after Rosas wins the Platine War. Events were added to properly notify ARC instead. An end_war command was added to prevent infinite war if someone is at war with ARC when it is annexed.