France releases Algeria and French Guiana
In default HFM, overseas departments such as Algeria and French Guiana frequently go unchanged when France is dismantled. This change adds release of state provinces in Algeria and French Guiana when France is dismantled, because it did not make sense for France to continue to possess overseas states and cores when other nations do not.