Constantine and North Central Sahara included
Constantine and North Central Sahara are usually colonies of France with Algerian cores, and in the case of Constantine eventually with French cores. The previous commits did not include these states in the dismantlement process, and France tended to retain them. Now any French cores are removed from these, and if colonial they are given out through the same process that awards colonies normally in dismantlement.