Colonial Conquest cb only usable against uncivilized
The `colonial_conquest` casus belli was usable against neighboring nations in Mozambique and Algeria, even if they are westernized. This seems to have been likely unintended, because it does not make much sense to perform a colonial conquest on a civilized nation. This has been patched to function like the Scramble For Africa casus belli, requiring the target to be civilized.