Core removal hopefully fixed using `THIS`
I discovered that this all did not work as intended because it requires `remove_core = THIS` instead of `remove_core = FRA` and changed the event accordingly. This was required because France is not only `FRA` but can instead be `BOR` or `VIC`.

During lengthy troubleshooting, I discovered that a `}` was used in a comment in the USA section, indicating a bracket imbalance that probably did not impact anything, but I changed it to a `)` anyway.