Convert outstanding files to cp1252
There were many UTF-8 encoded files still, this change converts them to
the cp1252 encoding that the game expects. For a couple files this only
changes some comments, but this also fixes some actual text bugs.

`PFH/localisation/0000_minimod.csv` is of note because it contained
UTF-8 mojibake that was not valid cp1252. Since the invalid key came
from base game files it was easy to restore it to the original
characters.