Heavenly Kingdom westernizes as Protestant
These changes were made because it was noticed that Heavenly Kingdom arbitrarily switches from Protestant to Mahayana upon westernization. This appears to have been a bug, and for the first commit I made some simple changes to keep it Protestant.

Religion specification upon westernization for main Chinese tags removed

There is no longer a religion specification for the main Chinese tags when they westernize in event `90900`. Additionally, it seemed best to keep the `protestant_country` and `change_to_protestant` country flags, because they are occasionally used by civilized tags.

In playtesting, I have guessed that pops will tend to be converted to the state religion if no other relevant flags are present. Here before westernization, pops in Heavenly Kingdom convert to Protestant as expected:
![HeavenlyKingdom](https://user-images.githubusercontent.com/17787203/69500522-d479b380-0eb0-11ea-9e96-905db6221d61.png)

After westernization, the conversion is still working:
![HeavenlyKingdom](https://user-images.githubusercontent.com/17787203/69500548-0ab73300-0eb1-11ea-91b4-41b32f8f4d08.png)

To see if the removal of specification of religion from the other tags (Qing Empire and Beiyang), I checked to see if conversion to Mahayana works as expected. Before westernization:
![HeavenlyKingdom](https://user-images.githubusercontent.com/17787203/69500616-7b5e4f80-0eb1-11ea-83b9-f3ef6379cb1c.png)

After westernization it is still working:
![HeavenlyKingdom](https://user-images.githubusercontent.com/17787203/69500644-b2346580-0eb1-11ea-8fd6-716e71255d27.png)

Commit--bugfix of prior changes, and streamlining for the 3 Chinese tags:

Prior upload had unintentional specification of religion, probably from an older copy of the China.txt file. New version removes specification of religion from that file. I retained different flag handling for `QNG`, `CHI`, and `TPG` because it seemed best to retain `protestant_country` and make sure that `change_to_protestant` is present for `TPG`. While `change_to_protestant` is probably not a massively important flag for `TPG`, it looked like it would be probably best to keep it. Line endings were checked to make sure that they are LF instead of CRLF.

Commit: quick update to add `set_country_flag = protestant_country` to `TPG`