Stop the religious effects in China more carefully
Closes #40.

When Qing loses the 2nd Opium War it may be forced to sign the Treaty of
Tientsin which historically allowed foreign missionary activity in China
among other things. To translate this into in-game mechanics, `QNG` is
temporarily switched to a protestant country. This is then undone upon
Westernisation, switching back to Mahayana. (There also are accompanying
country flags and modifiers.)

This has the unfortunate side-effect that Taiping also switches to
Mahayana upon Westernisation no matter what. With this change the
Westernisation of Chinese countries is more careful and preserves the
state religion of the Heavenly Kingdom.