Occupation of the Congo now with proper tech requirements, no longer spammed by ai, and no longer enabled when the effect cannot do anything
Commit: Catch file up to changes in master and fix encoding

Changes the Occupation of the Congo allow and potential to be equal, so that the potential does not show the decision as available before it is valid.

Commit: New fix that allows the decision to be spammed.

Determined that this decision is actually supposed to be spammed to "step up" from Gabon through Chad to occupy the Congo Basin. It was being spammed by the ai (and blocked other ai decisions) up until the invention of The Dark Continent. Therefore, changed it to require that invention before it would be enabled.

Commit: Fixing tech requirements to stop ai spam, other minor changes, fixing previous errors, and modernization of my work on this decision.

The `occupation_of_the_congo` decision is repaired and updated with proper cultural tech requirements for each iterative stage of the occupation. This fixes a problem where the decision would be active but ineffective until `the_dark_continent` was discovered, which had made the decision visible until then for nations that had the other requirements. The ai would also spam the decision until the effect could do something, at least slowing down the game. Some minor fixes were made to remove two whitespace errors, a potential encoding issue, and some ineffective lines in the `allow` statement. Requirements for the proper techs were added into each part of the `allow`, which already were required in their respective parts of the `effect`.

The problem looked like this:

![CongoOccupation](https://user-images.githubusercontent.com/17787203/75637182-968bda00-5bd9-11ea-8764-2f77159272a2.png)

and fixed by this and previous commits, it looks like this:

![CongoOccupation](https://user-images.githubusercontent.com/17787203/75640673-b37cd900-5bea-11ea-9202-0a75403d760a.png)

It was also tested for post-1880, pre-1890 functionality, and it seems to be fine:

![CongoOccupation](https://user-images.githubusercontent.com/17787203/75954165-e97ac100-5e67-11ea-94e3-70158f4f0559.png)

and:

![CongoOccupation](https://user-images.githubusercontent.com/17787203/75954188-f5668300-5e67-11ea-89f2-e3ce305a0706.png)

Commit: Fixing line endings and encoding (I hope), although github indicates errors in Lüderitzbucht and Maltahöhe, I think this may be an error in github's display. Otherwise, I need to find a new way of ensuring proper encoding.

Line endings were converted to CR LF by the Windows version of git, and therefore needed to be fixed. File changed to Windows-1252.