Normalise line endings
This gets rid of the stray '\r' carriage returns still hanging around.
As such this is best reviewed with `--ignore-cr-at-eol`.

The change was generated through the following clumsy zsh incantation:

    $ vim +':bufdo set ff=unix eol fixeol | %s/\n\r/\r/e | %s/\r\n/\r/e | %s/\r/\r/e | update | bdel' "${(@f)$(rg --sort=path --files-with-matches '\r' PFH)}"

Its effectiveness can be measured by the following ripgrep incantation:

    $ rg '\r' --files-with-matches PFH

To be sure that the change was harmless, special attention was directed
to the province files as some of them feature the most churn. In fact a
couple of these had nothing but carriage returns as file endings!
Inspection of some of these provinces while in-game show that all
information is correctly being read by the game.