Add safety valve system for dead diplomatic relationships
In some circumstances, code can start diplomatic relationships (e.g.
wars or overlord/vassal relationships) involving at least one
non-existing (i.e. 'dead') country. This is problematic when the game
cannot handle the situation, usually resulting in a relationship that
cannot be ended: non-existing countries cannot engage in diplomacy, nor
can existing countries send them diplomatic requests.

While any code that results in this situation is buggy and should be
fixed, sometimes things fall through the cracks. This change mitigates
the situation by adding a safety valve that automatically ends these
diplomatic relationships:

- wars
- overlord relationships (where the non-existing country controls the
  other)

Other relationships appear to be handled by the game just fine, or at
worst can be repaired after a save/reload. These two cases are the most
problematic as they respectively lock a country into permanent war or
permanent puppetdom, each with dire consequences.

(A related situation involves the 'ghost' armies or ships of a no longer
existing country. This is not tackled by this change, and should instead
involve porting the upstream fix.)

Due to game limitations, ending overlord relationships is a circuitous
process and involves transferring a lake(!!!) province to briefly revive
the dead country---and this takes until the following day to resolve
properly. Ideally a dedicated single-pixel sea or lake tile should be
used for the purpose, but this would involve modding the map and is left
to a future change.

Closes: #44, #84