Abandon .mod file versioning
This approach was considered for making it easier to switch around mod
versions, with implications on improved mod compatibility checks.
However the reality of having to copy settings around when jumping from
mod to mod already means that the user has to take manual steps.

As things stand, having the version in the .mod filename spares the user
very little effort and introduces the possibility for confusion.