World Fair planning fixed to become less spammy under certain conditions
While this likely occurs only under highly specific circumstances, the World's Fair planning event can become spammy for certain French primary tags such as Martinique.

This seems to be due to a modifier for tags with a French primary culture, but which do not have an Eiffel Tower built yet. The Eiffel Tower is linked to an 1889 World's Fair known as the Exposition Universelle, and the decision `build_the_eiffel_tower` requires the country flag `world_fair_planner`. This serves as evidence that the relatively extreme modifier is meant specifically for France, and not for any tags that can never build the Eiffel Tower.

I was previously puzzled over why `primary_culture` was used instead of `tag = FRA` in the event. Then I realized that France occasionally changes tags through the course of the game. Therefore, have adjusted the fix to scope France based on `primary_culture` and `capital`.

This is described in issue #87