Aden protectorate formation modified
References issues #59 and #36. Particularly, #59 lists historical references and discusses tuning decisions and localisation characterization for the added flavor.

The regular HFM seizing of Aden and formation of the Aden Protectorate are commented out to make the following changes, becuase it shared only a little with the new edits. Rather than changing the existing decisions, which would render them less recognizable in the files, I decided to make new files and add new decisions.

The problems addressed were originally noticed because of the lack of player protection for the owner of Aden, as detailed in #36. In discussion of this, historical clarification was provided by arkhometha, who also expressed a wish for the Aden Protectorate to be formed by diplomacy instead of the repeated wars that happened in HPM and HFM. This led to some additional research and work on my part to bring that about as a more elegant solution to the problem.

Historical flavor is added to portray the events leading to formation of the Aden Protectorate, with some minimum flavor included. It starts with flavor that can lead to the historical occupation of Aden by the United Kingdom. From there it describes the periodic Fadhli raids on Aden, and covers the 1866 British military retribution against Fadhli. An event is added that describes that Kathiri-Qu'aiti conflict that was recorded starting in 1858. For Mahra, the possibility of a Socotra purchase is added, by a Great Power owner of Bombay from certain continents, expected to be the United Kingdom. From there, if conditions are right, the Aden Protectorate will be formed. Informal protection is offered for Kathiri and Mahra before the Aden Protectorate is formed. Player protection exists for most of these stages, but excessively slow westernizers can lose by 1888. A decision is added for a late-released Kathiri or Mahra to petition to join Yemen if conditions are right. Added a brigade to Fadhli to properly reflect its belligerence, and especially to explain why Lahej would ask for help from the UK to fight Fadhli. Also I added a historical land military leader for Fadhli: Ahmad ibn `Abd Allah al-Fadli. New graphics are added for events and decisions. National Unification of Yemen no longer removes cores, because some internal instability is desired in Yemen.

Various notable bugfixes to the above:
Added an event for Fadhli Punishment, because it seemed inappropriate for the war start without explanation. Also, allowed it to occur even when Fadhli is in the sphere of the owner of Aden, because this seems allowable in historical terms. Some localisation was adjusted and added to make the text more flavorful.

Fixed an issue with tags not being offered informal protection if they were already in the great power's sphere.

Replaced modifiers for the Aden raids, since they appear to have no effect as event modifiers, according to the wiki. I replaced them with `import_cost` which hopefully has an effect.

Discovered that Mahra can refuse the Socotra Purchase while in the sphere of the country requesting it. This led me on a merry chase where I realized that `in_sphere = FROM` would not work, at least in this case, to specify `ai_chance` modifier factors. It instead led the Mahra ai to settle on one option out of three 100% of the time, even when its factor is 0. Eventually, I fixed the issue, and I have blamed it on a delay of 3 days being built in between the decision and the events that it spawns. This seemed to make the game confused over the tag that `FROM` references. I changed the delay to 0 days, and now I have my factors as I wanted them.

Added quotes to event localisation markers.

Fixed a bug with ai decision modifiers for tags that could petition to join Yemen, which had prevented the decision after it from working. Also removed the ai modifier factor that affected ai choice when Yemen had at least one brigade, because it did not work and was covered by another modifier.

Made the Kathiri-Qu'aiti conflict also happen to puppeted Yemen, as was historical.
Required Fadhli to have a brigade to raid Aden.

Made a minor change to localisation to inform the player of why the conflict happens to puppeted Yemen.

Fixes in response to review

Fixed various syntax errors. Prevented Fadhli from benefiting from raiding itself after conquering Aden. A syntax error was discovered and fixed in the informal protection decision. Added truce checks for anything starting a war. Cleaned up A Gift to Our Friends in Aden to work for any southern Yemeni owner of Aden. Added event for when Fadhli conquers Aden, so that it does not keep the Raiding Aden bonus. Adjusted the method of granting casus bellorum in events, so that they do not result in infamy. Added a check on the Informal Protection decision so that it is not visible after the Aden Protectorate is formed. Caused the Punishment of the Fadhli event to trigger the Raids on Aden Stop event, to notify players of the narrative and removal of the modifier. Adjusted Mahra ai chances for the Socotra Purchase, because I discovered that it was almost impossible for Mahra to be sphered by 1845. Adjusted the Fadhli Raids Stop event to fire only once. Changed Informal Protection so that it requires the target to not be in the sphere of the country that uses it. Deleted previous decisions that are replaced by new content. Added a civilization check for decisions that check for progress towards civilization. Fixed ai factors to calculate correctly. Added war checks for all Yemeni tags that had a chance of being annexed. Added a territorial requirement for Yemen formation, so that an exiled tag could not be used to trigger it.

Alliance breaking of Fadhli on Fadhli punishment now only applies to Nejd.

At long last, worked out how to prevent a great power (usually the UK) from giving away Socotra to an allied Yemeni tag after previously purchasing it.

Added a requirement to the Deria Dowlat event, meant for an uncivilized owner of Aden, to make sure that they are not at war before the event can fire.

Changed Fadhli's role in the Informal Protection decision, because Fadhli was often escaping being sphered due to losing its army during the Saudi Civil War. Now, when Fadhli loses its army it will not break ai Yemen formation, since that does not seem to be a reasonable way for events to unfold.

Fixed syntax error with tag.

Gave event images a better background so that they blend with the event window. Made a minor localisation change to make a sentence flow better.

Added an event to supply more suitable flavor text for a non-Yemeni owner of Aden who sees the Fadhli Raids event.

Added Zaydi flavor to the formation of Yemen, and coordinated timing of informal protection events to follow completion of the Suez Canal.

A great power (usually the UK) now offers informal protection only if the Ottoman Empire puppets or annexes Zaydi. Zaydi gains a harsh modifier while it is a puppet of the Ottomans, to simulate historical unrest. Also overhauled informal protection itself to permit variation in timing, and adjusted the decision picture to make it more similar to other decision pictures in exposure and color balance. Also added the Greater Yemen decision, with expanded claims. It is worth consideration whether too many provinces (from Muscat to all of Hejaz) gain cores, this can be reduced--I did this because I prefer not to add cores to some but not all provinces of a state.

Fixed a typo in localisation

I added a declaration of war on the Ottoman decision to puppet Zaydi, because the Ottomans were very rarely using their granted casus belli. It also seemed better to require Zaydi to not belong to a sphere before the Ottomans would use a decision to wage war to puppet them. Required that the Ottomans not be at war or disarmed before attacking Zaydi