Prevents core removal for existing tags in most cases
References #70 and #73. In many cases, the cores of tags that start uncivilized are removed by decisions and events that are meant to simulate historical change from uncivlized areas to colonies to post-colonial areas. However, if cores are considered to be claims, this should not occur while the tag still exists. This is different from a situation where tags give up cores in treaties, when this fix should not be applied. Additionally, I have allowed core removal in cases where the tag no longer exists, so that the simulation occurs as intended. I have ignored removal of ENG cores in the New World, since those should happen always through ENG giving up cores voluntarily. In other cases, especially within Europe, it seemed like a bad idea to change anything.

In the course of making changes, I removed `cleanup_colonial_cores`, since that decision should not exist if things function as intended.

I changed `organize_ghana` and `reorganize_ghana` to require a tag to own the in-game capital of Ashanti, since core removals were not working while Ashanti and other tags with claims on Ghana existed. This seemed justifiable to me since otherwise only a single coastal province was required to organize all of Ghana.

I removed a few apparently arbitrary Malay core changes in `97122`, `Anglo-Siamese Treaty of 1909`, because it did not make sense to switch around and not remove the claims of tags that do not exist. I additionally made some changes to add Malaysia cores properly to its provinces.

The initial set of changes succeeded with only a few noticed bugs. These and matters requiring investigation are listed and discussed [in this post](https://github.com/moretrim/ccHFM/issues/70#issuecomment-1023876986).

Found an overlooked core removal of potentially existing Kurdish tags.

Acadia core removal now removes the cores on Maritime Union territory as well. Louisiana core removal now covers the possibility of USA core removal.